chore(admin): convert useDebounce to TS
