--- packages/core/admin/admin/src/hooks/useDebounce/index.js
@@ -1,19 +1,0 @@
-import { useEffect, useState } from 'react';
-
-const useDebounce = (value, delay) => {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-};
-
-export default useDebounce;
