+++ packages/core/admin/admin/src/pages/MarketplacePage/tests/plugins.test.js
@@ -15,7 +15,9 @@ import server from './server';
 // Increase the jest timeout to accommodate long running tests
 jest.setTimeout(50000);
 jest.mock('../hooks/useNavigatorOnline');
-jest.mock('../../../hooks/useDebounce', () => (value) => value);
+jest.mock('../../../hooks/useDebounce', () => ({
+  useDebounce: jest.fn((value) => value),
+}));
 jest.mock('@strapi/helper-plugin', () => ({
   ...jest.requireActual('@strapi/helper-plugin'),
   useAppInfo: jest.fn(() => ({
